CONFIGDB.TST: Update basic interface tests.
